Reorganize into GithubReckoner model instead
Following the pattern of the SamAccountReckoner
